Fix Lexical floating toolbar with conditional root element
