+++ packages/liveblocks-react-lexical/src/use-root-element.ts
@@ -1,0 +1,19 @@
+import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
+import { useCallback, useSyncExternalStore } from "react";
+
+export function useRootElement(): HTMLElement | null {
+  const [editor] = useLexicalComposerContext();
+
+  const subscribe = useCallback(
+    (onStoreChange: () => void) => {
+      return editor.registerRootListener(onStoreChange);
+    },
+    [editor]
+  );
+
+  const getSnapshot = useCallback(() => {
+    return editor.getRootElement();
+  }, [editor]);
+
+  return useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
+}
